Use the Spring Initializr
I rewrote the document and updated the code (mostly changing package names) to use the Spring Initializr.
