+++ complete/src/main/java/com/example/schedulingtasks/SchedulingTasksApplication.java
@@ -1,0 +1,14 @@
+package com.example.schedulingtasks;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@SpringBootApplication
+@EnableScheduling
+public class SchedulingTasksApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SchedulingTasksApplication.class);
+	}
+}
