+++ complete/src/test/java/com/example/schedulingtasks/SchedulingTasksApplicationTest.java
@@ -14,25 +14,27 @@
  * limitations under the License.
  */
 
-package hello;
+package com.example.schedulingtasks;
 
-import java.text.SimpleDateFormat;
-import java.util.Date;
+import static org.assertj.core.api.Assertions.assertThat;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.stereotype.Component;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
 
-@Component
-public class ScheduledTasks {
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class SchedulingTasksApplicationTest {
 
-	private static final Logger log = LoggerFactory.getLogger(ScheduledTasks.class);
+	@Autowired
+	private ScheduledTasks tasks;
 
-	private static final SimpleDateFormat dateFormat = new SimpleDateFormat("HH:mm:ss");
-
-	@Scheduled(fixedRate = 5000)
-	public void reportCurrentTime() {
-		log.info("The time is now {}", dateFormat.format(new Date()));
+	@Test
+	public void contextLoads() {
+		// Basic integration test that shows the context starts up properly
+		assertThat(tasks).isNotNull();
 	}
+
 }
