+++ initial/src/main/java/com/example/schedulingtasks/SchedulingTasksApplication.java
@@ -1,14 +1,13 @@
-package hello;
+package com.example.schedulingtasks;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
-@EnableScheduling
-public class Application {
+public class SchedulingTasksApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(Application.class);
+		SpringApplication.run(SchedulingTasksApplication.class, args);
 	}
+
 }
