+++ initial/src/test/java/com/example/schedulingtasks/SchedulingTasksApplicationTests.java
@@ -1,0 +1,16 @@
+package com.example.schedulingtasks;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class SchedulingTasksApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
